# Slightly simplify the lookup of data in `Dict.{get, getAsync, has}`